requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.12.16
requests==2.31.0
